chore(gux-popover-list): replace gux-popover-list with gux-popover-list-beta
