--- packages/genesys-spark-components/src/components/beta/gux-popover-list-beta/example.html
@@ -1,58 +1,0 @@
-<h1>gux-popover-list-beta</h1>
-<div class="scroll-container">
-  <div class="scroll-interior">
-    <div id="popover-target">Example Element</div>
-    <gux-popover-list-beta
-      id="popover-example"
-      position="top"
-      for="popover-target"
-      is-open
-    >
-      <gux-list>
-        <gux-list-item onclick="notify(event, 'item-1')">Item 1</gux-list-item>
-        <gux-list-item onclick="notify(event, 'item-2')">Item 2</gux-list-item>
-        <gux-list-item onclick="notify(event, 'item-3')">Item 3</gux-list-item>
-        <gux-list-item onclick="notify(event, 'item-4')">Item 4</gux-list-item>
-        <gux-list-item onclick="notify(event, 'item-5')">Item 5</gux-list-item>
-        <gux-list-item onclick="notify(event, 'item-6')">Item 6</gux-list-item>
-      </gux-list>
-    </gux-popover-list-beta>
-  </div>
-</div>
-
-<style
-  onload="(function () {
-  const popoverTarget = document.getElementById('popover-target');
-  const popoverExample = document.getElementById('popover-example');
-
-  popoverTarget.addEventListener('mouseover', () => {
-    popoverExample.isOpen = true;
-  });
-})()"
->
-  .scroll-container {
-    width: 900px;
-    height: 300px;
-    margin: auto;
-    overflow: scroll;
-    border: 1px solid #aaa;
-  }
-
-  .scroll-interior {
-    position: relative;
-    width: 600px;
-    height: 600px;
-    background-color: #fff;
-  }
-
-  #popover-target {
-    position: absolute;
-    top: 250px;
-    left: 250px;
-    display: inline-block;
-    padding: 10px 15px;
-    background-color: #fdfdfd;
-    border: 1px #e4e9f0;
-    box-shadow: 0 0 2px fade(@gux-black-30, 24%);
-  }
-</style>
