+++ packages/genesys-spark-components/src/components/stable/gux-dismiss-button/readme.md
@@ -21,7 +21,6 @@ This component is meant for use in other components as a dismiss button.
  - [gux-popover](../gux-popover)
  - [gux-popover-beta](../../beta/gux-popover-beta)
  - [gux-popover-list](../gux-popover-list)
- - [gux-popover-list-beta](../../beta/gux-popover-list-beta)
  - [gux-simple-toast-legacy](../../legacy/gux-simple-toast)
  - [gux-toast](../gux-toast)
 
@@ -38,7 +37,6 @@ graph TD;
   gux-popover --> gux-dismiss-button
   gux-popover-beta --> gux-dismiss-button
   gux-popover-list --> gux-dismiss-button
-  gux-popover-list-beta --> gux-dismiss-button
   gux-simple-toast-legacy --> gux-dismiss-button
   gux-toast --> gux-dismiss-button
   style gux-dismiss-button fill:#f9f,stroke:#333,stroke-width:4px
