+++ packages/genesys-spark-components/src/components/stable/gux-popover-list/example.html
@@ -2,7 +2,12 @@
 <div class="scroll-container">
   <div class="scroll-interior">
     <div id="popover-target">Example Element</div>
-    <gux-popover-list id="popover-example" position="top" for="popover-target">
+    <gux-popover-list
+      id="popover-example"
+      position="top"
+      for="popover-target"
+      is-open
+    >
       <gux-list>
         <gux-list-item onclick="notify(event, 'item-1')">Item 1</gux-list-item>
         <gux-list-item onclick="notify(event, 'item-2')">Item 2</gux-list-item>
@@ -19,8 +24,9 @@
   onload="(function () {
   const popoverTarget = document.getElementById('popover-target');
   const popoverExample = document.getElementById('popover-example');
+
   popoverTarget.addEventListener('mouseover', () => {
-    popoverExample.hidden = false;
+    popoverExample.isOpen = true;
   });
 })()"
 >
